# pkg/standart-gorm/internal/utils/http/response.go
package http

import (
	"encoding/json"
	"log"
	"net/http"
)

type ResponseError struct {
	Status int       `json:"status"`
	Result errorData `json:"result"`
}

type errorData struct {
	Error string `json:"error"`
}

type response struct {
	Status int         `json:"status"`
	Result interface{} `json:"result"`
}

func newResponse(data interface{}, status int) *response {
	return &response{
		Status: status,
		Result: data,
	}
}

func (resp *response) bytes() []byte {
	data, _ := json.Marshal(resp)
	return data
}

func (resp *response) sendResponse(w http.ResponseWriter, _ *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	w.Header().Set("Access-Control-Allow-Origin", "*")
	w.WriteHeader(resp.Status)
	_, _ = w.Write(resp.bytes())
	log.Println(resp.bytes())
}

// 200
func StatusOK(w http.ResponseWriter, r *http.Request, data interface{}) {
	newResponse(data, http.StatusOK).sendResponse(w, r)
}

// 204
func StatusNoContent(w http.ResponseWriter, r *http.Request) {
	newResponse(nil, http.StatusNoContent).sendResponse(w, r)
}

// 400
func StatusBadRequest(w http.ResponseWriter, r *http.Request, err error) {
	data := errorData{Error: err.Error()}
	newResponse(data, http.StatusBadRequest).sendResponse(w, r)
}

// 404
func StatusNotfound(w http.ResponseWriter, r *http.Request, err error) {
	data := errorData{Error: err.Error()}
	newResponse(data, http.StatusNotFound).sendResponse(w, r)
}

// 405
func StatusMethodNotAllowed(w http.ResponseWriter, r *http.Request) {
	newResponse(nil, http.StatusMethodNotAllowed).sendResponse(w, r)
}

// 500
func StatusInternalServerError(w http.ResponseWriter, r *http.Request, err error) {
	data := errorData{Error: err.Error()}
	newResponse(data, http.StatusInternalServerError).sendResponse(w, r)
}